% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xy_from_lonlat.R
\name{convert_lonlat}
\alias{convert_lonlat}
\title{Convert data.frame from lon/lat to UTM planar coordinates}
\usage{
convert_lonlat(df, epsg_utm)
}
\arguments{
\item{df}{A data.frame containing \code{lon} and \code{lat} columns}

\item{epsg_utm}{EPSG code of the UTM projection}
}
\value{
Data.frame with added \code{x} and \code{y} columns (meters)
}
\description{
Convert data.frame from lon/lat to UTM planar coordinates
}
\keyword{internal}
