% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rect_stand.R
\name{create_rect_stand}
\alias{create_rect_stand}
\title{Center the surveyed trees of the core polygon in a rectangle plot}
\usage{
create_rect_stand(
  trees,
  cell_size,
  core_polygon_df = NULL,
  use_rect_zone = FALSE,
  fill_around = FALSE
)
}
\arguments{
\item{trees}{a data.frame with one row for each tree and their position in a (x,y) grid.
It must be composed of two columns "x" and "y"}

\item{cell_size}{double - Length of the side of a squared cell composing the stand (in meters)}

\item{core_polygon_df}{data.frame - Coordinates of the vertices composing the core polygon.
It must be composed of two columns "x" and "y"}

\item{use_rect_zone}{boolean - How to consider the delimitation of the inventory zone,
used to define the square plot and when fill_around is TRUE, for filling the trees aorund.
Either to use the given or computed core polygon (FALSE),
or the minimum-area enclosing rectangle from the polygon (TRUE).}

\item{fill_around}{boolean - If TRUE, add trees around the core polygon within the stand
until the stand reach the same total basal area per hectare than the core polygon}
}
\description{
Center the surveyed trees of the core polygon in a rectangle plot
}
