% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_cloture20}
\alias{data_cloture20}
\title{Information for cloture20 inventory}
\format{
\code{data_cloture20}, a named list with 5 elements:
\describe{
\item{trees}{data.frame that contains for each tree, the information about
its location, species, size, size of the crown and crown information.
\describe{
\item{id_tree}{Unique id of the tree. (integer)}
\item{species}{Species Latin name. (character)}
\item{x, y}{Coordinates of the base of the tree in the forest stand in m. (double)}
\item{dbh_cm}{Diameter at breast height (1.30m) of the trunk of the tree in cm. (double)}
\item{crown_type}{Type of the crown between paraboloid (P) and ellispoid (E). (character)}
\item{h_m}{Height of the tree trunk in m. (double)}
\item{hbase_m}{Crown base height of the tree (i.e. height at wich the crown start) in m. (double)}
\item{hmax_m}{Height at which the crown radius is maximum.
Set it to NA in case of simple crown shapes ("P" and "E"), otherwise, your values won't be considered as hmax is automatically computed.}
\item{rn_m, rs_m, re_m, rw_m}{Biggest radius of the tree crown in m (double)}
\item{crown_openess}{Crown Openness of the tree (no unit) (i.e. Fraction of the energy of a light ray crossing the crown that is intercepted).
Used when computing interception with a crown considered as a porous envelop. (double)}
\item{crown_lad}{Leaf Area Density of the tree crown in m2/m3 (i.e. surface of leave per volume of crown, considering an homogeneous crown).
Used when computing interception with a crown considered as a turbid medium. (double)}
}
}
\item{species_color}{named vector that contains color HEX code for each species}
\item{sensors}{data.frame that contains information about the stand sensors}
\describe{
\item{id_sensor}{Unique id of the sensor (integer)}
\item{x, y}{Coordinates of the sensor in the stand (double)}
\item{h_m}{Height of the sensor in m. (double)}
\item{pacl, pacl_direct, pacl_diffuse}{Measured proportion of above canopy light on the sensor,
for respectively total, direct and diffuse energies}
}
\item{core_polygon}{data.frame containing vertices of the tree inventory zone, described by the coordinates (x, y) of each vertex edge.}
\item{radiation}{data.frame that contains monthly radiations.
\describe{
\item{month}{Unique id of the tree. (integer)}
\item{Hrad}{Monthly radiation in a horizontal plane in MJ. (double)}
\item{DGratio}{Ratio between monthly diffuse and global energies. (double)}
}
}
\item{info}{A named numeric vector with site information:
\describe{
\item{latitude}{Latitude of the site (decimal degrees).}
\item{longitude}{Longitude of the site (decimal degrees).}
\item{size_x}{Horizontal extent of the site (meters). Here, it is set to NA as the inventory is defined by the core polygon.}
\item{size_y}{Vertical extent of the site (meters). Here, it is set to NA as the inventory is defined by the core polygon.}
\item{slope}{Mean slope of the terrain (degrees).}
\item{aspect}{Aspect (degrees from north).}
\item{north_to_x_cw}{Angle from north to the x-axis (clockwise, degrees).}
}
}
}
}
\source{
Gauthier Ligot (Gembloux Agro-Bio Tech)
}
\usage{
data(data_cloture20)
}
\description{
Information about cloture20 inventory in Belgium used
as an example for a complex core polygon and demonstrating light sensors.
}
\keyword{datasets}
