% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_prenovel}
\alias{data_prenovel}
\title{Information for Prenovel inventory}
\format{
\code{data_prenovel}, a named list with 3 elements:
\describe{
\item{trees}{data.frame that contains for each tree, the information about
its location, species, size, size of the crown and crown information.
\describe{
\item{id_tree}{Unique id of the tree. (integer)}
\item{species}{Species Latin name. (character)}
\item{x, y}{Coordinates of the base of the tree in the forest stand in m. (double)}
\item{dbh_cm}{Diameter at breast height (1.30m) of the trunk of the tree in cm. (double)}
\item{crown_type}{Type of the crown between paraboloid (P) and ellispoid (E). (character)}
\item{h_m}{Height of the tree trunk in m. (double)}
\item{hbase_m}{Crown base height of the tree (i.e. height at wich the crown start) in m. (double)}
\item{hmax_m}{Height at which the crown radius is maximum.
Set it to NA in case of simple crown shapes ("P" and "E"), otherwise, your values won't be considered as hmax is automatically computed.}
\item{rn_m, rs_m, re_m, rw_m}{Biggest radius of the tree crown in m (double)}
\item{crown_openess}{Crown Openness of the tree (no unit) (i.e. Fraction of the energy of a light ray crossing the crown that is intercepted).
Used when computing interception with a crown considered as a porous envelop. (double)}
\item{crown_lad}{Leaf Area Density of the tree crown in m2/m3 (i.e. surface of leave per volume of crown, considering an homogeneous crown).
Used when computing interception with a crown considered as a turbid medium. (double)}
}
}
\item{species_color}{named vector that contains color HEX code for each species}
\item{radiation}{data.frame that contains monthly radiations.
\describe{
\item{month}{Unique id of the tree. (integer)}
\item{Hrad}{Monthly radiation in a horizontal plane in MJ. (double)}
\item{DGratio}{Ratio between monthly diffuse and global energies. (double)}
}
}
\item{info}{A named numeric vector with site information:
\describe{
\item{latitude}{Latitude of the site (decimal degrees).}
\item{longitude}{Longitude of the site (decimal degrees).}
\item{size_x}{Horizontal extent of the site (meters).}
\item{size_y}{Vertical extent of the site (meters).}
\item{slope}{Mean slope of the terrain (degrees).}
\item{aspect}{Aspect (degrees from north).}
\item{north_to_x_cw}{Angle from north to the x-axis (clockwise, degrees).}
}
}
}
}
\source{
Courbaud Benoit (INRAe LESSEM Grenoble)
}
\usage{
data(data_prenovel)
}
\description{
Information about Prenovel marteloscope in Jura (France) used
as an example for light interception by symetrical crowns.
}
\keyword{datasets}
