% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_rad_bechefa}
\alias{data_rad_bechefa}
\title{Radiation dataset for Bechefa (Belgium)}
\format{
\code{data_rad_bechefa}
A data.frame with 12 rows and 3 columns:
\describe{
\item{month}{Unique id of the tree. (integer)}
\item{Hrad}{Monthly radiation in a horizontal plane in MJ. (double)}
\item{DGratio}{Ratio between monthly diffuse and global energies. (double)}
}
}
\source{
PVGIS : https://joint-research-centre.ec.europa.eu/pvgis-photovoltaic-geographical-information-system_en
}
\usage{
data_rad_bechefa
}
\description{
Dataset with monthly radiation of Bechefa forest in Belgium.
It is useful for runnning SamsaRaLight on Bechefa stand using \code{SamsaRaLight::data_trees_bechefa} dataset.
Data were fecteched from PVGIS using the function \code{SamsaRaLight::get_monthly_rad()} at
latitude 50.04 and longitude 5.2
}
\keyword{datasets}
