% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bottom_azimut}
\alias{get_bottom_azimut}
\title{Compute bottom azimut}
\usage{
get_bottom_azimut(aspect, north2x)
}
\arguments{
\item{aspect}{double - Angle of slope bottom on the compass from the North, clockwise rotation (in degrees)
northern aspect : 0, eastern aspect : 90, southern aspect : 180, western aspect : 270}

\item{north2x}{double - Angle from North to x axis clockwise. (in degrees)
Default correspond to a Y axis oriented toward the North.}
}
\value{
a bottom azimuth numeric value
}
\description{
Compute bottom azimut
}
\keyword{internals}
