% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_intercept.R
\name{get_intercept_coords}
\alias{get_intercept_coords}
\title{Compute the coordinates of the interception point from the distance to
the target cell and beam direction.}
\usage{
get_intercept_coords(
  length,
  cos_elevation,
  sin_elevation,
  cos_azimut,
  sin_azimut
)
}
\arguments{
\item{length}{\itemize{
\item Distance from intersection point and target center cell (in meters)
}}

\item{cos_elevation}{\itemize{
\item Cosinus of elevation angle of the beam (in radians)
}}

\item{sin_elevation}{\itemize{
\item Sinus of elevation angle of the beam (in radians)
}}

\item{cos_azimut}{\itemize{
\item Cosinus of azimut angle of the beam (in radians)
}}

\item{sin_azimut}{\itemize{
\item Sinus of azimut angle of the beam (in radians)
}}
}
\value{
a vector containing the coordinates of the interception point and
its distance to the target cell
}
\description{
Compute the coordinates of the interception point from the distance to
the target cell and beam direction.
}
