% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_rad.R
\name{get_monthly_rad}
\alias{get_monthly_rad}
\title{Fetch monthly radiations}
\source{
https://joint-research-centre.ec.europa.eu/pvgis-photovoltaic-geographical-information-system_en
}
\usage{
get_monthly_rad(
  latitude,
  longitude,
  start_year = 2005,
  end_year = 2020,
  average_years = TRUE
)
}
\arguments{
\item{latitude}{latitude of the plot}

\item{longitude}{longitude of the plot}

\item{start_year}{positive integer between 2005 and 2020 - start year on which to fetch monthly data}

\item{end_year}{positive integer between 2005 and 2020 - end year on which to fetch monthly data}

\item{average_years}{if TRUE, years are averaged to obtain a single value for each month}
}
\value{
Monthly horizontal radiation (Hrad) and diffuse to global ratio (DGratio)
averaged between start_year and end_year
}
\description{
Fetch monthly radiation data from PVGIS website (by API)
between start and end year (limit years are from 2005 to 2020).
Fetched variables are Hrad = horizontal plane irradiation and
DGratio = ratio of diffuse to global radiation (in horizontal plane).

! YOU NEED AN INTERNET CONNECTION TO ACCESS THE DATA BY API !
}
