% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xy_from_lonlat.R
\name{get_utm_epsg}
\alias{get_utm_epsg}
\title{Get EPSG code for UTM zone from longitude and latitude}
\usage{
get_utm_epsg(lon, lat)
}
\arguments{
\item{lon}{Mean longitude (decimal degrees)}

\item{lat}{Mean latitude (decimal degrees)}
}
\value{
EPSG code (numeric)
}
\description{
Get EPSG code for UTM zone from longitude and latitude
}
\keyword{internal}
