% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_z}
\alias{get_z}
\title{Compute z coordinate of a point (x,y).}
\usage{
get_z(x, y, slope_rad, bottom_azimut_rad)
}
\arguments{
\item{x}{X-coordinate of the point}

\item{y}{Y-coordinate of the point}

\item{slope_rad}{Slope of the stand (in radians)}

\item{bottom_azimut_rad}{Azimuth of the vector orthogonal to the ground in the x,y system (in radians)}
}
\value{
a z coordinate numeric value
}
\description{
Compute z coordinate of a point (x,y).
}
\keyword{internals}
