% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sl_stand.R
\name{plot.sl_stand}
\alias{plot.sl_stand}
\title{Plot a SamsaRaLight virtual stand}
\usage{
\method{plot}{sl_stand}(x, ..., top_down = FALSE, only_inv = FALSE)
}
\arguments{
\item{x}{An object of class \code{sl_stand}.}

\item{...}{Additional arguments passed to lower-level plotting functions.}

\item{top_down}{Logical, if TRUE, creates a top-down view with multiple directions (south, north, west, east).}

\item{only_inv}{Logical, if TRUE, plot only trees from the initial inventory (i.e. not trees added to fill around the core polygon)}
}
\value{
A \code{ggplot} object representing the stand.
}
\description{
This function plots a virtual forest stand (\code{sl_stand}) produced by SamsaRaLight.
It can display a top-down view with tree crowns or a side/top view with cells and trees.
}
