% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sl_output.R
\name{plot_sl_output}
\alias{plot_sl_output}
\title{Function to plot SamsaRaLight object output}
\usage{
plot_sl_output(
  sl_output,
  trees.only_inventoried = FALSE,
  trees.border.species = FALSE,
  trees.fill = "e",
  trees.fill.palette = c("viridis", "light", "base"),
  trees.fill.inverse = FALSE,
  trees.fill.limits = NULL,
  cells.border = FALSE,
  cells.fill = "e",
  cells.fill.palette = c("light", "viridis", "base"),
  cells.fill.inverse = FALSE,
  cells.fill.limits = NULL,
  sensors.plot = FALSE
)
}
\description{
@param plot.trees Character string indicating the filling variables of the trees.
Set it to NULL if you do not want to plot the trees
}
