% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stand_3D.R
\name{plot_stand_3D}
\alias{plot_stand_3D}
\title{Function to plot SamsaRaLight stand in 3D}
\usage{
plot_stand_3D(
  trees = NULL,
  species_colors = NULL,
  slope = 0,
  aspect = 0,
  north_to_x_cw = 90
)
}
\arguments{
\item{trees}{a tree table}
}
\description{
Function to plot SamsaRaLight stand in 3D
}
