% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rotate_vec_ccw}
\alias{rotate_vec_ccw}
\title{Rotate 2D vectors by an angle (counter-clockwise)}
\usage{
rotate_vec_ccw(x, y, theta)
}
\arguments{
\item{x}{Numeric vector of x coordinates.}

\item{y}{Numeric vector of y coordinates.}

\item{theta}{Rotation angle in radians (counter-clockwise).}
}
\value{
A data.frame with rotated coordinates:
\describe{
\item{x}{Rotated x coordinates}
\item{y}{Rotated y coordinates}
}
}
\description{
Internal utility to rotate 2D coordinates by a given angle (in radians),
using a standard rotation matrix.
}
\details{
The rotation follows:
\deqn{
x' = x \cos(\theta) - y \sin(\theta)
}
\deqn{
y' = x \sin(\theta) + y \cos(\theta)
}
}
\keyword{internal}
