% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamsaRaLight-package.R
\docType{package}
\name{SamsaRaLight-package}
\alias{SamsaRaLight}
\alias{SamsaRaLight-package}
\title{SamsaRaLight: Forest light interception and radiation simulation}
\description{
SamsaRaLight is an R package for simulating forest stand light interception,
energy balance and radiation transfer using spatially explicit tree crowns,
sensors and grid cells.
}
\details{
The package provides tools to:
\itemize{
\item Build forest stands with trees, cells and sensors
\item Run ray-tracing based light simulations
\item Compute tree, cell and sensor energy variables
\item Summarize and visualize simulation outputs
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://natheob.github.io/SamsaRaLight/}
  \item Report bugs at \url{https://github.com/natheob/SamsaRaLight/issues/}
}

}
\author{
\strong{Maintainer}: Natheo Beauchamp \email{beauchamp.natheo@gmail.com} (\href{https://orcid.org/0009-0007-9103-5194}{ORCID}) (R package developer)

Authors:
\itemize{
  \item Gauthier Ligot \email{gligot@uliege.be} (\href{https://orcid.org/0000-0002-5508-4358}{ORCID}) (Java Capsis developer and algorithm improvement)
  \item Francois de Coligny \email{francois.decoligny@inrae.fr} (\href{https://orcid.org/0000-0002-8538-3009}{ORCID}) (Java Capsis developer)
  \item Maxime Jaunatre \email{maxime.jaunatre@inrae.fr} (\href{https://orcid.org/0009-0002-2816-1677}{ORCID}) (R package support developer)
  \item Benoit Courbaud \email{benoit.courbaud@inrae.fr} (\href{https://orcid.org/0000-0002-3050-9559}{ORCID}) (Algorithm and theory creator) [copyright holder]
}

}
\keyword{internal}
