% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_model.R
\name{sl_beerlambert}
\alias{sl_beerlambert}
\title{Compute intercepted energy using Beer Lambert law}
\usage{
sl_beerlambert(
  incident_energy,
  path_length,
  leaf_area_density = 0.5,
  extinction_coef = 0.5,
  clumping_factor = 1
)
}
\arguments{
\item{incident_energy}{energy that arrive to the crown}

\item{path_length}{length of the path within the crown (in meters)}

\item{leaf_area_density}{LAD of the crown (m2 of leaves per m3 of crown volume)}

\item{extinction_coef}{Probability of a leaf to intercept the ray (linked to leaf orientation)}

\item{clumping_factor}{Aggregation of leaves within the crown volume (1 is homogeneous)}
}
\description{
Compute intercepted energy using Beer Lambert law
}
