% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_model.R
\name{sl_compute_energy_porous}
\alias{sl_compute_energy_porous}
\title{Compute potential and intercepted energy from interceptions (rays X trees)
using porous envelop (cf Ligot et al. 10.1139/cjfr-2013-0494)}
\usage{
sl_compute_energy_porous(
  interceptions,
  slope_rad,
  bottom_azimut_rad,
  cell_surface
)
}
\description{
Compute potential and intercepted energy from interceptions (rays X trees)
using porous envelop (cf Ligot et al. 10.1139/cjfr-2013-0494)
}
\keyword{internal}
