% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sl_rays.R
\name{sl_compute_nrj_diffuse}
\alias{sl_compute_nrj_diffuse}
\title{Compute energy of diffuse ray}
\usage{
sl_compute_nrj_diffuse(
  soc,
  total_diffuse,
  height_angle_rad,
  diffuse_anglestep_rad
)
}
\arguments{
\item{soc}{boolean - Standard Overcast Sky, if false: Uniform Overcast Sky}

\item{total_diffuse}{double - Total diffuse energy on a horizontal plan (in kWh.m-2)}

\item{height_angle_rad}{double - Angle between beam and soil (in radians)}

\item{diffuse_anglestep_rad}{double - Hour angle between two diffuse beams (in radians)}
}
\value{
Energy per square meter of a horizontal plan (in MJ.m-2)
}
\description{
Calculating SamsaraLight rays diffuse Energy in MJ/m2 of a plane
perpendicular to beam ray direction for a Standard Overcast Sky or
Uniform Overcast Sky are possible
}
\keyword{internal}
