% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_rays.R
\name{sl_create_monthly_rays}
\alias{sl_create_monthly_rays}
\title{Compute direct and diffuse rays of a growing season}
\usage{
sl_create_monthly_rays(
  monthly_rad,
  latitude,
  start_day = 1,
  end_day = 365,
  soc = TRUE,
  slope = 0,
  north_to_x_cw = 90,
  aspect = 0,
  height_anglemin = 10,
  direct_startoffset = 0,
  direct_anglestep = 5,
  diffuse_anglestep = 15
)
}
\arguments{
\item{monthly_rad}{data.frame - Monthly horizontal radiation (Hrad) and diffuse to global ratio (DGratio)
Computed with function samsaRa::sl_get_monthlyrad()}

\item{latitude}{double - Latitude of the plot (in degrees)}

\item{start_day}{integer between 1 and 365 - First day of the vegetative period}

\item{end_day}{integer between 1 and 365 - Last day of the vegetative period}

\item{soc}{boolean - Standard Overcast Sky, if false: Uniform Overcast Sky}

\item{slope}{double - Slope of the plot (in degrees)}

\item{north_to_x_cw}{double - Angle from North to x axis clockwise. (in degrees)
Default correspond to a Y axis oriented toward the North.}

\item{aspect}{double - Angle of slope bottom on the compass from the North, clockwise rotation (in degrees)
northern aspect : 0, eastern aspect : 90, southern aspect : 180, western aspect : 270}

\item{height_anglemin}{double - Angle minimum between beam and soil (in degrees)}

\item{direct_startoffset}{double - Angle at which to start first direct ray (in degrees)}

\item{direct_anglestep}{double - Hour angle between two direct beams (in degrees)}

\item{diffuse_anglestep}{double - Hour angle between two diffuse beams (in degrees)}
}
\value{
list of 3 elements : horizontal energy (double), slope energy (double)
and rays (data.frame) with n rows and 5 columns:
\itemize{
\item{azimut}{Azimut of the ray in radians}
\item{height_angle}{Angle between beam and soil (in radians)}
\item{e}{Energy of ray before crossing the canopy (in MJ.m-2)}
\item{direct}{true if the ray is direct false if it is diffuse}
}
}
\description{
Compute direct and diffuse rays of a growing season
}
