% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sl_rays.R
\name{sl_create_rays_diffuse}
\alias{sl_create_rays_diffuse}
\title{Create diffuse rays}
\usage{
sl_create_rays_diffuse(
  soc,
  total_diffuse,
  height_anglemin_rad,
  diffuse_anglestep_rad,
  slope_rad,
  bottom_azimut_rad
)
}
\arguments{
\item{soc}{boolean - Standard Overcast Sky, if false: Uniform Overcast Sky}

\item{total_diffuse}{double - Total diffuse energy on a horizontal plan (in kWh.m-2)}

\item{height_anglemin_rad}{double - Angle minimum between beam and soil (in radians)}

\item{diffuse_anglestep_rad}{double - Hour angle between two diffuse beams (in radians)}

\item{slope_rad}{double - Slope of the plan (in radians)}

\item{bottom_azimut_rad}{double - Azimut of the vector orthogonal to the ground
in the (x,y) system (in radians)}
}
\value{
list of 3 elements : horizontal energy (double), slope energy (double)
and diffuse rays (data.frame)
}
\description{
Create SamsaraLight diffuse rays of a plane with a classical sky
hemisphere divided by meridians and parallels
can use Standard Overcast Sky or Uniform Overcast Sky
}
\keyword{internal}
