% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_model.R
\name{sl_get_interceptions}
\alias{sl_get_interceptions}
\title{Search for interceptions between all rays coming to all target cells and trees}
\usage{
sl_get_interceptions(potcells_rays, trees)
}
\description{
Search for interceptions between all rays coming to all target cells and trees
}
