% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_model.R
\name{sl_get_potentialcells_rays_relative}
\alias{sl_get_potentialcells_rays_relative}
\title{Get relative cell ids that could contain trees potentially intercepting a given ray}
\usage{
sl_get_potentialcells_rays_relative(
  rays,
  cell_size,
  slope_rad,
  bottom_azimut_rad,
  max_height,
  max_cradius
)
}
\arguments{
\item{rays}{data.frame with n rows and 5 columns:
\itemize{
\item{id}{Unique id of the ray}
\item{azimut}{Azimut of the ray in radians}
\item{height_angle}{Angle between beam and soil (in radians)}
\item{e_init}{Initital energy of ray before crossing the canopy (in MJ.m-2)}
\item{e_current}{Current energy of ray (in MJ.m-2)}
\item{direct}{true if the ray is direct false if it is diffuse}
}}

\item{cell_size}{Length of the side of a squared cell composing the stand (in meters)}

\item{slope_rad}{Slope of the stand (in radians)}

\item{bottom_azimut_rad}{Azimut of the vector orthogonal to the ground in the x,y system (in radians)}

\item{max_height}{Max possible height of a tree in Samsara model (in meters)}

\item{max_cradius}{Max possible crown radius of a tree in Samsara model (in meters)}
}
\value{
Return a data.frame with n rows for each ray X possible cell and 3 columns
\itemize{
\item{id_ray}{Unique id of the ray}
\item{xid_rel}{Relative X-axis id compared to the target cell (cell where ray is targeted to its center)}
\item{yid_rel}{Relative Y-axis id compared to the target cell (cell where ray is targeted to its center)}
}
}
\description{
The trees which can intercept a beam are located in cells with
their center located in a competition rectangle of length L (beam X direction)
\itemize{
\item R (opposite direction) and width R (directions perpendicular X to beam).
}
}
