% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_intercept.R
\name{sl_intercept_crown_8thellipsoidal}
\alias{sl_intercept_crown_8thellipsoidal}
\title{Compute ray interception with a 8th of ellipsoidal crown}
\usage{
sl_intercept_crown_8thellipsoidal(
  elevation,
  azimut,
  x,
  y,
  z,
  x_shift,
  y_shift,
  z_shift,
  a,
  b,
  c
)
}
\arguments{
\item{elevation}{double - Elevation angle of the ray (in radians)}

\item{azimut}{double - azimut angle of the ray (in radians, trigonometric 0 and counter clockwise)}

\item{x}{double - X-component of ellipsoid center (in meters)}

\item{y}{double - Y-component of ellipsoid center (in meters)}

\item{z}{double - Z-component of ellipsoid center (in meters)}

\item{x_shift}{double - Shift of ellipsoid center on X-axis (in meters)}

\item{y_shift}{double - Shift of ellipsoid center on Y-axis (in meters)}

\item{z_shift}{double - Shift of ellipsoid center on Z-axis (in meters)}

\item{a}{double - Parameter a of ellipsoid semi-principal axes (X-axis, semi-major)}

\item{b}{double - Parameter b of ellipsoid semi-principal axes (Y-axis, semi-minor)}

\item{c}{double - Parameter c of ellipsoid semi-principal axes (Z-axis, height)}
}
\value{
Returns the interception path length and a distance to
the origin (i.e. the target cell center) (NA is no interception)
}
\description{
Evaluates if the given ray intercepts this crown, previously relocated
with the given shifts.
}
