% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_intercept.R
\name{sl_intercept_crown_paraboloid}
\alias{sl_intercept_crown_paraboloid}
\title{Compute ray interception with an ellipsoidal crown}
\usage{
sl_intercept_crown_paraboloid(
  cos_elevation,
  sin_elevation,
  cos_azimut,
  sin_azimut,
  x,
  y,
  z,
  x_shift,
  y_shift,
  z_shift,
  a,
  b,
  h
)
}
\arguments{
\item{cos_elevation}{double - Cosinus ofelevation angle of the ray (in radians)}

\item{sin_elevation}{double - Sinus ofelevation angle of the ray (in radians)}

\item{cos_azimut}{double - Cosinus of azimut angle of the ray (in radians, trigonometric 0 and counter clockwise)}

\item{sin_azimut}{double - Sinus of azimut angle of the ray (in radians, trigonometric 0 and counter clockwise)}

\item{x}{double - X-component of paraboloid center (in meters)}

\item{y}{double - Y-component of paraboloid center (in meters)}

\item{z}{double - Paraboloid base height (in meters)}

\item{x_shift}{double - Shift of paraboloid center on X-axis (in meters)}

\item{y_shift}{double - Shift of paraboloid center on Y-axis (in meters)}

\item{z_shift}{double - Shift of paraboloid center on Z-axis (in meters)}

\item{a}{double - Parameter a of paraboloid = semi-axis of basal ellipse along X axis}

\item{b}{double - Parameter b of paraboloid = semi-axis of basal ellipse along Y axis}

\item{h}{double - Parameter h of paraboloid = crown height, along Z axis}
}
\value{
Returns the interception path length and a distance to
the origin (i.e. the target cell center) (NA is no interception)
}
\description{
Evaluates if the given ray intercepts this crown, previously relocated
with the given shifts.
}
\keyword{internal}
