% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_run.R
\name{sl_run}
\alias{sl_run}
\title{Compute samsara ligth radiative balance}
\usage{
sl_run(
  trees,
  monthly_rad,
  latitude = 46,
  start_day = 1,
  end_day = 365,
  slope = 0,
  north_to_x_cw = 90,
  aspect = 0,
  cell_size = 10,
  n_cells = 10,
  use_torus = TRUE,
  turbid_medium = TRUE,
  use_rcpp = TRUE
)
}
\arguments{
\item{trees}{a data.frame with one row for each tree and 9 columns:
\itemize{
\item{"id": }{Unique id of the tree}
\item{"x": }{X position of the tree within the stand}
\item{"y": }{Y position of the tree within the stand}
\item{"z": }{Z position of the tree within the stand}
\item{"height": }{Height of the trees (in meters)}
\item{"cradius": }{Crown radius of the tree (in meters)}
\item{"cdepth": }{Crown depth (i.e. height of the crown) of the tree (in meters)}
\item{"crown_lad"}{Leaf Area Density (in m2 of leaves per m3 of crown)}
\item{"crown_p"}{Crown openness (no unit)}
\item{"crown_type"}{Between "P" for paraboloid and "E" for ellipsoid}
\item{"id_cell"}{Unique id of the cell the tree belong to (i.e. \code{cells} data.frame parameter)}
}}

\item{monthly_rad}{data.frame - Monthly horizontal radiation (Hrad) and diffuse to global ratio (DGratio)
Computed with function samsaRa::sl_get_monthlyrad()}

\item{latitude}{double - Latitude of the plot (in degrees)}

\item{start_day}{integer between 1 and 365 - First day of the vegetative period}

\item{end_day}{integer between 1 and 365 - Last day of the vegetative period}

\item{slope}{double - Slope of the plot (in degrees)}

\item{north_to_x_cw}{double - Angle from North to x axis clockwise. (in degrees)
Default correspond to a Y axis oriented toward the North.}

\item{aspect}{double - Angle of slope bottom on the compass from the North, clockwise rotation (in degrees)
northern aspect : 0, eastern aspect : 90, southern aspect : 180, western aspect : 270}

\item{cell_size}{Length of the side of a squared cell composing the stand (in meters)}

\item{n_cells}{integer - Number of cells of the side of a squared stand}

\item{use_torus}{if True, compute light competition using borders modelled with a
torus system, otherwise,borders are open grasslands}

\item{turbid_medium}{If TRUE, crown are considered as turbid medium, otherwise,
considered as porous envelope}

\item{use_rcpp}{If TRUE, use Rcpp package and C++ implementation of SamsaraLight for faster
computation. Otherwise, run SamsaraLight with only R-based scripts}
}
\description{
Compute samsara ligth radiative balance
}
