% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_run_capsis.R
\name{sl_run_capsis}
\alias{sl_run_capsis}
\title{Run SamsaraLight from Capsis}
\usage{
sl_run_capsis(
  capsis_folderpath,
  inv_fp,
  use_rcapsis = TRUE,
  java_server_started = FALSE,
  commandfile_fp = NULL
)
}
\description{
Run SamsaraLight from Capsis
}
