% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_model.R
\name{sl_summarize_interceptions}
\alias{sl_summarize_interceptions}
\title{Summarize interceptions output into trees and cells interception datasets}
\usage{
sl_summarize_interceptions(
  interceptions,
  trees_id,
  cells,
  cell_surface_slope,
  e_slope_tot
)
}
\arguments{
\item{trees_id}{vector of ids of the trees composing the stand}

\item{cells}{data.frame with id_cell and corresponding cell center coordinates}

\item{e_slope_tot}{Sum of diffuse and direct energy coming into a cell considering slope (MJ)}

\item{out_targets}{Data.frame with bound targets output from sl_run_target()}

\item{cell_surface}{Area of a cell considering slope (in m2)}
}
\value{
a list of trees and cells interception datasets
}
\description{
Summarize interceptions output into trees and cells interception datasets
}
