% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sl_stand.R
\name{summary.sl_stand}
\alias{summary.sl_stand}
\title{Summary of a SamsaRaLight stand}
\usage{
\method{summary}{sl_stand}(object, ...)
}
\arguments{
\item{object}{A \code{sl_stand} object}

\item{...}{Unused}
}
\value{
Invisibly returns a list of summary tables
}
\description{
Summary of a SamsaRaLight stand
}
