% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sl_advanced.R
\name{validate_interception_model}
\alias{validate_interception_model}
\title{Validate tree interception model parameters against model type}
\usage{
validate_interception_model(trees, turbid_medium)
}
\description{
Validate tree interception model parameters against model type
}
\keyword{internal}
