% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_polygon.R
\name{check_polygon}
\alias{check_polygon}
\title{Check and validate a polygon defined by vertices}
\usage{
check_polygon(polygon_df, trees_inv, sensors = NULL, verbose = TRUE)
}
\arguments{
\item{polygon_df}{A data.frame with columns x and y defining polygon vertices}

\item{trees_inv}{A data.frame with one row per tree.
See \link{check_inventory} for the required structure and validated columns.}

\item{sensors}{Optional data.frame defining position and height of the sensor within the stand.
See \link{check_sensors} for the required structure and validated columns.}

\item{verbose}{Logical. If TRUE, warnings are printed}
}
\value{
A data.frame of polygon vertices (x, y):
\itemize{
\item unchanged if valid
\item modified if fixed (with a warning)
}
}
\description{
This function converts a data.frame of polygon vertices into an sf POLYGON
and checks its validity. If the polygon is invalid, it attempts to fix it.
}
