% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xy_from_lonlat.R
\name{create_xy_from_lonlat}
\alias{create_xy_from_lonlat}
\title{Create planar (x, y) coordinates from longitude / latitude}
\usage{
create_xy_from_lonlat(df)
}
\arguments{
\item{df}{A data.frame containing geographic coordinates (\code{lon} and \code{lat}).}
}
\value{
A list with the following elements:
\describe{
\item{\code{df}}{The input data.frame with added \code{x} and \code{y} columns (meters, UTM).}
\item{\code{epsg}}{EPSG code of the UTM projection used.}
}
}
\description{
This function converts geographic coordinates (\code{lon}, \code{lat})
into planar coordinates (\code{x}, \code{y})
using an automatically selected UTM projection.
}
\details{
The input data.frame must contain geographic coordinates (\code{lon} and \code{lat}).
Planar coordinates (\code{x}, \code{y}) are automatically computed using the
UTM zone inferred from the mean longitude and hemisphere inferred from the mean latitude.
}
