% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sl_output.R
\name{plot.sl_output}
\alias{plot.sl_output}
\title{Plot a SamsaRaLight output}
\usage{
\method{plot}{sl_output}(
  x,
  ...,
  what_trees = c("compet", "intercepted", "potential"),
  what_cells = c("relative", "absolute"),
  show_trees = TRUE,
  direct_energy = NULL
)
}
\arguments{
\item{x}{An object of class \code{sl_output}, returned by \code{run_sl()}.}

\item{...}{Additional arguments passed to lower-level plotting functions.}

\item{what_trees}{Character; which tree metric to plot. Choices are:
\describe{
\item{"compet"}{Light competition index (LCI), reversed viridis scale.}
\item{"intercepted"}{Intercepted energy (MJ).}
\item{"potential"}{Potential intercepted energy (MJ).}
}
Default is "compet".}

\item{what_cells}{Character; which cell (ground) metric to plot. Choices are:
\describe{
\item{"relative"}{Proportion of above canopy light (PACL)).}
\item{"absolute"}{Energy on the ground (MJ/m2).}
}
Default is "relative".}

\item{show_trees}{Logical; whether to display trees on top of the ground light map. Default is TRUE.}

\item{direct_energy}{Logical or NULL.
If NULL (default), total radiation outputs are plotted (direct + diffuse).
If TRUE, only direct radiation components are plotted.
If FALSE, only diffuse radiation components are plotted.
This option requires \code{detailed_output = TRUE} when running the simulation}
}
\value{
A ggplot object.
}
\description{
Visualize ground light and tree energy metrics for a \code{sl_output} object.
}
