% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inventory.R
\name{plot_inventory}
\alias{plot_inventory}
\title{Plot a from-above view of a tree inventory}
\usage{
plot_inventory(trees_inv, show_id = TRUE)
}
\arguments{
\item{trees_inv}{A data.frame of trees that passed \link{check_inventory}.}

\item{show_id}{Logical; if TRUE (default), displays tree identifiers at crown centers.}
}
\value{
A ggplot object displaying the trees as ellipses in a from-above view.
}
\description{
Visualizes a forest inventory as ellipses representing tree crowns.
}
