% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sl_output.R
\name{plot_sl_output}
\alias{plot_sl_output}
\title{@param plot.trees Character string indicating the filling variables of the trees.
Set it to NULL if you do not want to plot the trees}
\usage{
plot_sl_output(
  sl_output,
  trees.border.species = FALSE,
  trees.fill = "species",
  trees.fill.inverse = FALSE,
  cells.border = FALSE,
  cells.fill = NULL,
  cells.fill.palette = c("base", "base01", "viridis", "viridis01", "light", "light01")
)
}
\description{
@param plot.trees Character string indicating the filling variables of the trees.
Set it to NULL if you do not want to plot the trees
}
