% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stand_2D.R
\name{plot_stand_2D}
\alias{plot_stand_2D}
\title{Function to plot SamsaRaLight trees with top/down view}
\usage{
plot_stand_2D(sl_stand, top_down = FALSE, transparency = TRUE)
}
\arguments{
\item{trees}{a tree table}
}
\description{
Function to plot SamsaRaLight trees with top/down view
}
