% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamsaRaLight-package.R
\docType{package}
\name{SamsaRaLight-package}
\alias{SamsaRaLight}
\alias{SamsaRaLight-package}
\title{SamsaRaLight: Forest light interception and radiation simulation}
\description{
SamsaRaLight is an R package for simulating forest stand light interception,
energy balance and radiation transfer using spatially explicit tree crowns,
sensors and grid cells.
}
\details{
The package provides tools to:
\itemize{
\item Build forest stands with trees, cells and sensors
\item Run ray-tracing based light simulations
\item Compute tree, cell and sensor energy variables
\item Summarize and visualize simulation outputs
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://natheob.github.io/SamsaRaLight/}
}

}
\author{
\strong{Maintainer}: Natheo Beauchamp \email{beauchamp.natheo@gmail.com} (R package developper)

Authors:
\itemize{
  \item Gauthier Ligot \email{gligot@uliege.be} (Java Capsis developper and algorithm improvment) [copyright holder]
  \item Francois de Coligny \email{coligny@cirad.fr} (Java Capsis developper)
  \item Maxime Jaunatre \email{maxime.jaunatre@inrae.fr} (R package support developper)
  \item Benoit Courbaud \email{benoit.courbaud@inrae.fr} (Algorithm and theory creator) [copyright holder]
}

}
\keyword{internal}
