% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sl_rays.R
\name{sl_compute_sunazimut}
\alias{sl_compute_sunazimut}
\title{Compute sun azimut}
\usage{
sl_compute_sunazimut(
  latitude_rad,
  declination_rad,
  hour_angle_rad,
  height_angle_rad,
  southazimut_ccw_rad
)
}
\arguments{
\item{latitude_rad}{double - Latitude of the plot (in radians)}

\item{declination_rad}{double - Declination angle in radians:
angle between the equator and a line drawn from the centre of the Earth to
the centre of the sun}

\item{hour_angle_rad}{double - Angle which the sun moves across the sky (in radians)}

\item{height_angle_rad}{double - Angle between beam and soil (in radians)}

\item{southazimut_ccw_rad}{double - Azimuth of south counterclockwise from x axis
in the (x,y) system (in radians)}
}
\description{
Computaion of sun azimut for a given height angle reference system with
angle origin on X > 0 axis and trigonometric rotation
}
\keyword{internal}
