% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_run.R
\name{sl_run}
\alias{sl_run}
\title{Compute samsara ligth radiative balance}
\usage{
sl_run(
  trees,
  monthly_rad,
  sensors = NULL,
  sensors_only = FALSE,
  latitude = 46,
  start_day = 1,
  end_day = 365,
  slope = 0,
  north_to_x_cw = 90,
  aspect = 0,
  cell_size = 10,
  n_cells_x = 10,
  n_cells_y = 10,
  soc = TRUE,
  height_anglemin = 10,
  direct_startoffset = 0,
  direct_anglestep = 5,
  diffuse_anglestep = 15,
  detailed_output = FALSE,
  use_torus = TRUE,
  turbid_medium = TRUE,
  trunk_interception = FALSE
)
}
\arguments{
\item{trees}{a data.frame with one row for each tree and 14 columns describing tree composing the stand:
\itemize{
\item{id: }{Unique id of the tree}
\item{x: }{X position of the tree within the stand}
\item{y: }{Y position of the tree within the stand}
\item{dbh_cm: }{Diameter at breast height (1.3m) of the tree trunk (in cm)}
\item{crown_type: }{
\itemize{
\item{"E": symetric ellispoidal crown with the maximum crown radius at the middle of the crown depth.
Crown radius is the mean value of the radius towards the fourth cardinal points}
\item{"P": symetric paraboloidal crown with the maximum crown radius at the crown base.
crown radius is the mean value of the radius towards the fourth cardinal points}
\item{"2E": ellipsoidal crown composed of two semi-ellipsoid (one above and one below).
Same as "E" but height of each semi-ellipsoid is given by hmax_m variable}
\item{"8E": ellispoidal crown composed of eight eighth of ellipsoids (four above and four below).
Same as "2E" but radius of each eight of ellipsoid is given by rn_m, rs_m, re_m, rw_m variables.}
\item{"4P": paraboloidal crown composed of four fourth of paraboloid.
Maximum crown radius at the crown base and radius of each fourth of paraboloid is given by rn_m, rs_m, re_m, rw_m variables.}
}
}
\item{h_m: }{Height of the tree (in meters)}
\item{hbase_m: }{Height of the base of the tree crown (in meters)}
\item{hmax_m: }{Height of the maximum radius of the tree crown (in meters)}
\item{rn_m: }{Radius of the tree crown toward Northern direction (in meters)}
\item{rs_m: }{Radius of the tree crown toward Southern direction (in meters)}
\item{re_m: }{Radius of the tree crown toward Eastern direction (in meters)}
\item{rw_m: }{Radius of the tree crown toward Western direction (in meters)}
\item{crown_openess: }{Crown openess of the crown (no unit) when considering porous envelop
(i.e. constant proportion of light energy interception when a ray intercepts a crown)}
\item{crown_lad: }{Leaf Area Density (in m2 of leaves per m3 of crown) when considering turbid medium
(i.e. density parameter of Beer Lambert law when a ray intercepts a crown)}
}}

\item{monthly_rad}{data.frame - Monthly horizontal radiation (Hrad) and diffuse to global ratio (DGratio)
Computed with function \code{samsaRaLight::sl_get_monthlyrad()}}

\item{sensors}{data.frame - Position of the sensor within the stand. Can be set to NULL if no sensor.
\itemize{
\item{id: }{Unique id of the sensor}
\item{x: }{X position of the sensor within the stand}
\item{y: }{Y position of the sensor within the stand}
\item{h_m: }{Height above ground of the sensor (in meters)}
}}

\item{sensors_only}{boolean - To compute interception only for the sensors ?}

\item{latitude}{double - Latitude of the plot (in degrees)}

\item{start_day}{integer between 1 and 365 - First day of the vegetative period}

\item{end_day}{integer between 1 and 365 - Last day of the vegetative period}

\item{slope}{double - Slope of the plot (in degrees)}

\item{north_to_x_cw}{double - Angle from North to x axis clockwise. (in degrees)
Default correspond to a Y axis oriented toward the North.}

\item{aspect}{double - Angle of slope bottom on the compass from the North, clockwise rotation (in degrees)
northern aspect : 0, eastern aspect : 90, southern aspect : 180, western aspect : 270}

\item{cell_size}{double - Length of the side of a squared cell composing the stand (in meters)}

\item{n_cells_x}{integer - Number of cells of the X-axis side of the rectangle stand}

\item{n_cells_y}{integer - Number of cells of the Y-axis side of the rectangle stand}

\item{soc}{boolean - Standard Overcast Sky, if false: Uniform Overcast Sky}

\item{height_anglemin}{double - Angle minimum between beam and soil (in degrees)}

\item{direct_startoffset}{double - Angle at which to start first direct ray (in degrees)}

\item{direct_anglestep}{double - Hour angle between two direct beams (in degrees)}

\item{diffuse_anglestep}{double - Hour angle between two diffuse beams (in degrees)}

\item{detailed_output}{boolean - If TRUE, sensors/cells/trees outputs also contain diffuse/direct energies,
and sensors/cells outputs contain energies both on the slope and on a horizontal plane.
If FALSE, energy given is only the total energy (sum of diffuse and direct) on the slope for trees and cells
and on a horizontal plane for sensors.}

\item{use_torus}{if True, compute light competition using borders modelled with a
torus system, otherwise,borders are open grasslands}

\item{turbid_medium}{If TRUE, crown are considered as turbid medium, otherwise,
considered as porous envelope}

\item{trunk_interception}{Consider interception of rays by trunks}
}
\description{
Compute samsara ligth radiative balance
}
