% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sl_output.R
\name{summary.sl_output}
\alias{summary.sl_output}
\title{Summary of a SamsaRaLight simulation}
\usage{
\method{summary}{sl_output}(object, ...)
}
\arguments{
\item{object}{an object of class \code{sl_output}}

\item{...}{unused}
}
\description{
Summary of a SamsaRaLight simulation
}
